// src/app/shared/interfaces/sobre-nos.ts
export interface SobreNos {
    id: number
    descricao: string
    endereco: string
    coordenadorLaboratorioId: number
    emailCoordenador: string
    telefoneLaboratorio: string
    segundaLaboratorioAbre: boolean
    horarioSegundaAbertura: string
    horarioSegundaFechamento: string
    tercaLaboratorioAbre: boolean
    horarioTercaAbertura: string
    horarioTercaFechamento: string
    quartaLaboratorioAbre: boolean
    horarioQuartaAbertura: string
    horarioQuartaFechamento: string
    quintaLaboratorioAbre: boolean
    horarioQuintaAbertura: string
    horarioQuintaFechamento: string
    sextaLaboratorioAbre: boolean
    horarioSextaAbertura: string
    horarioSextaFechamento: string
    sabadoLaboratorioAbre: boolean
    horarioSabadoAbertura: string
    horarioSabadoFechamento: string
    domingoLaboratorioAbre: boolean
    horarioDomingoAbertura: string
    horarioDomingoFechamento: string
    situacaoCadastroId: number
    createdAt: string
    userCreatedId: number
    updatedAt: string
    userUpdatedId: number
  }
  